extract FactLineBuilder to new file
